Crate libudev

Structs

Attribute

A device attribute.

Attributes

Iterator over a device’s attributes.

Context

A libudev context. Contexts may not be sent or shared between threads. The libudev(3) manpage says:

Device

A structure that provides access to sysfs/kernel devices.

Devices

Iterator over devices.

Enumerator

An enumeration context.

Error

The error type for libudev operations.

Event

An event that indicates a change in device state.

Monitor

Monitors for device events.

MonitorSocket

An active monitor that can receive events.

Properties

Iterator over a device’s properties.

Property

A device property.

Enums

ErrorKind

Types of errors that occur in libudev.

EventType

Types of events that can be received from udev.

Type Aliases

Result

A Result type for libudev operations.