Struct libudev::MonitorSocket

pub struct MonitorSocket<'a> {
    // some fields omitted
}

An active monitor that can receive events.

The events received by a MonitorSocket match the filters setup by the Monitor that created the socket.

Monitors are initially setup to receive events from the kernel via a nonblocking socket. A variant of poll() should be used on the file descriptor returned by the AsRawFd trait to wait for new events.

Methods

impl<'a> MonitorSocket<'a>

fn receive_event<'b>(&'b mut self) -> Option<Event<'a>>

Receives the next available event from the monitor.

This method does not block. If no events are available, it returns None immediately.